pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "quant"
version = "0.2.0"
description = "Internal automated trading system for Kalshi crypto binary markets."
readme = "README.md"
requires-python = ">=3.11"
keywords = ["kalshi", "trading", "kelly-criterion"]
dependencies = [
    "aiohttp>=3.9",
    "anthropic>=0.40",
    "cryptography>=42.0",
    "python-dotenv>=1.0.0",
    "websockets>=14.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff"]

[project.scripts]
quant = "quant.tools.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["quant"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
